root_number/requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
django-celery-results==2.0.0
django-cors-headers==3.6.0
djangorestframework==3.12.2
flower==0.9.5
future==0.18.2
humanize==3.2.0
isort==5.6.4
kombu==4.6.11
prometheus-client==0.8.0
prompt-toolkit==3.0.8
pycodestyle==2.6.0
pytz==2020.4
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==1.3.0
wcwidth==0.2.5
